backend_fastapi/pyproject.toml
[project]
name = "sso-gcp-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "authlib>=1.6.0",
    "fastapi[all]>=0.115.14",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
]
